requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-bootstrap4==1.1.1
django-el-pagination==3.3.0
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.8
uWSGI==2.0.18
